[refactor] add classes and create start.py script
