+++ ActiveLearning.py
@@ -1,11 +1,8 @@
 import logging
 import os
-import platform
 import time
-from tkinter import Tk
-import tkinter
 import pyautogui
-from typing import Iterable, Optional, Tuple, Union
+from typing import Iterable, Union
 from modAL.uncertainty import uncertainty_sampling
 import sklearn
 import numpy as np
@@ -13,6 +10,8 @@ from datasets import load_dataset
 from Annotation import Annotation
 from AnnotationFile import AnnotationFile
 from Dataset import Dataset
+from Gui import GUI
+from System import System
 from TextFile import TextFile
 
 from constants import FOLDER_NAME, PATH_TO_BRAT, COLLECTION_NAME, SUGGESTION_ANNOTATION_TYPE
@@ -24,11 +23,14 @@ class ActiveLearning:
     """
 
     def __init__(self):
-        logging.basicConfig(filename='example.log', filemode='w', encoding='utf-8', level=logging.INFO)
+        pass
 
-    def iteration(self, classifier: sklearn.base.BaseEstimator,
+    def     iteration(self, classifier: sklearn.base.BaseEstimator,
                   unlabeled_data: Union[list, np.ndarray],
                   num_to_annotate: int = 1):
+        gui = GUI()
+        system = System()
+
         indices = uncertainty_sampling(classifier=classifier, 
                                        X=unlabeled_data, 
                                        n_instances=num_to_annotate)
@@ -41,20 +43,20 @@ class ActiveLearning:
 
         title = "Suggestions loaded"
         message = "Suggestions has been loaded.\nYou can now start annotating."
-        self.show_custom_popup(title, message)
+        gui.show_custom_popup(title, message)
         pyautogui.hotkey('f5')
 
-        path_to_collection, file_names = self.get_file_names_from_path(path_to_brat=PATH_TO_BRAT, folder_name=FOLDER_NAME, collection_name=COLLECTION_NAME)
+        path_to_collection, file_names = system.get_file_names_from_path(path_to_brat=PATH_TO_BRAT, folder_name=FOLDER_NAME, collection_name=COLLECTION_NAME)
         
         while self.suggestions_left_in_files(path=path_to_collection, file_names=file_names):
             self.check_file_change(path=path_to_collection, file_names=file_names)
         logging.info("Annotation by domain expert finished. No suggestions left.")
         title = "Annotation finished"
         message = "You finished the current annotation step.\nNow the next training iteration began.\nPlease do not change any file until the next call."
-        self.show_custom_popup(title, message)
+        gui.show_custom_popup(title, message)
 
         dataset = Dataset(path_to_collection=path_to_collection)
-        storage_path = "../data/" #must be changed after testing
+        storage_path = "./data/" 
         file_name = "training_dataset.json"
         dataset.to_json(storage_path, file_name)
         logging.info(f"Updated dataset with new annotations is generated and saved under {storage_path + file_name}")
@@ -77,7 +79,8 @@ class ActiveLearning:
         ---------
             samples (Iterable[str]): Samples to add to annotation files
         """
-        path_to_collection, file_names = self.get_file_names_from_path(path_to_brat=PATH_TO_BRAT, folder_name=FOLDER_NAME, collection_name=COLLECTION_NAME)
+        system = System()
+        path_to_collection, file_names = system.get_file_names_from_path(path_to_brat=PATH_TO_BRAT, folder_name=FOLDER_NAME, collection_name=COLLECTION_NAME)
 
         text_files = [
             TextFile(file_name=file_name, path=path_to_collection)
@@ -121,51 +124,6 @@ class ActiveLearning:
                 current_mod_time[idx] = os.path.getmtime(path + file_name)
         
         return True
-    
-    def show_custom_popup(self, titel: str, message: str) -> None:
-        """
-        Displays a custom popup window with the given title and message.
-
-        Arguments
-        ---------
-            title (str): The title of the popup window.
-            message (str): The message to be displayed in the popup window.
-        """
-        popup = Tk()
-        popup.title(titel)
-        popup.eval('tk::PlaceWindow . center')
-        popup.geometry("300x120")
-        label = tkinter.Label(popup, text=message, padx=20, pady=20, anchor="w")
-        label.pack()
-
-        ok_button = tkinter.Button(popup, text="OK", command=popup.destroy, width=20)
-        ok_button.pack(pady=10)
-
-        popup.mainloop()
-
-    def get_file_names_from_path(self, path_to_brat: str, folder_name: str, collection_name: Optional[str]) -> Tuple[str, list[str]]:
-        """
-        Retrieves the file names from a specified path in the file system.
-
-        Arguments
-        ---------
-            path_to_brat (str): The base path to the BRAT directory.
-            folder_name (str): The name of the folder within the BRAT directory.
-            collection_name (str): The name of the collection within the folder (optional).
-
-        Returns
-        -------
-            Tuple[str, List[str]]: A tuple containing the collection path and a list of file names.
-                                The collection path is formed by concatenating the base path, folder name,
-                                and collection name (if provided). The list of file names contains the names
-                                of the files within the collection path.
-        """
-        operating_system = platform.system()
-        slash = "\\" if operating_system == "Windows" else "/"
-        collection_path = path_to_brat + slash + folder_name + slash
-        collection_path += collection_name + slash if collection_name else ""
-
-        return collection_path, os.listdir(collection_path)
     
     def suggestions_left_in_files(self, path: str, file_names: list[str]) -> bool:
         """
