+++ constants.py
@@ -1,7 +1,7 @@
 
 SUGGESTION_ANNOTATION_TYPE = "TO_BE_ANNOTATED"
 
-PATH_TO_BRAT = ".."
+PATH_TO_BRAT = "."
 FOLDER_NAME = "doc"
 COLLECTION_NAME = None
 FILE_NAME = None
